+++ src/main/java/application/controller/AbstractController.java
@@ -1,4 +1,4 @@
-package webserver.controller;
+package application.controller;
 
 import org.springframework.http.HttpMethod;
 import webserver.domain.HttpRequest;
@@ -6,13 +6,10 @@ import webserver.domain.HttpResponse;
 
 public class AbstractController implements Controller {
 
-    protected boolean login = false;
-
     @Override
     public void service(HttpRequest httpRequest, HttpResponse httpResponse) {
         if (httpRequest.getHeader("Cookie").contains("logined=true")) {
             httpResponse.addHeader("Set-Cookie", "logined=true");
-            login = true;
         }
 
         if (httpRequest.getMethod().equals(HttpMethod.POST)) {
