+++ src/main/java/webserver/RequestHandler.java
@@ -1,8 +1,8 @@
 package webserver;
 
+import application.controller.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.controller.*;
 import webserver.domain.HttpRequest;
 import webserver.domain.HttpResponse;
